+++ msd_baja_ecvt_ecu/Core/Inc/print_debug.h
@@ -8,7 +8,7 @@
 #ifndef INC_PRINT_DEBUG_H_
 #define INC_PRINT_DEBUG_H_
 #include "stdint.h"
-void print_debug_sensor_vals(uint32_t prim_rpm, uint32_t sec_rpm, uint16_t throttle_angle, uint16_t helix_angle);
+void print_debug_sensor_vals(float prim_rpm, float sec_rpm, float throttle_angle, float helix_angle);
 
 
 #endif /* INC_PRINT_DEBUG_H_ */
