+++ msd_baja_ecvt_ecu/Core/Src/print_debug.c
@@ -8,11 +8,11 @@
 #include <stdio.h>
 #include <stdint.h>
 
-void print_debug_sensor_vals(uint32_t prim_rpm, uint32_t sec_rpm, uint16_t throttle_angle, uint16_t helix_angle){
+void print_debug_sensor_vals(float prim_rpm, float sec_rpm, float throttle_angle, float helix_angle){
 	printf(
-			"Primary RPM: %10u\n\r"
-			"Secondary RPM: %10u\n\r"
-			"Throttle Angle: %4u\n\r"
-			"Helix Angle: %4u\n\r"
+			"Primary RPM: %6.3f\n\r"
+			"Secondary RPM: %6.3f\n\r"
+			"Throttle Angle: %2.3f\n\r"
+			"Helix Angle: %2.3f\n\r"
 			"\x1B[1;1H", prim_rpm, sec_rpm, throttle_angle, helix_angle);
 }
